requirements.txt
fastapi==0.104.0
SQLAlchemy==2.0.22
uvicorn==0.23.2
python-dotenv==1.0.0
typer==0.3.2
pydantic-settings==2.0.3